requirements.txt
numpy>=1.19.0
pandas>=1.1.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
scipy>=1.5.0
